breaking: move navigation, form and page types to their $app modules
