+++ .changeset/tidy-waves.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': major
+---
+
+breaking: move navigation, form and page types to `$app/navigation`, `$app/forms` and `$app/state`
