+++ packages/kit/src/exports/public.d.ts
@@ -24,11 +24,7 @@ import { BuildData, SSRNodeLoader, SSRRoute, ValidatedConfig } from 'types';
 import { SvelteConfig } from '@sveltejs/vite-plugin-svelte';
 import { StandardSchemaV1 } from '@standard-schema/spec';
 import { Plugin } from 'vite';
-import {
-	RouteId as AppRouteId,
-	LayoutParams as AppLayoutParams,
-	ResolvedPathname
-} from '$app/types';
+import { RouteId as AppRouteId, LayoutParams as AppLayoutParams } from '$app/types';
 
 export { PrerenderOption } from '../types/private.js';
 
@@ -1205,294 +1201,6 @@ export interface NavigationEvent<
 }
 
 /**
- * Information about the target of a specific navigation.
- */
-export interface NavigationTarget<
-	Params extends AppLayoutParams<'/'> = AppLayoutParams<'/'>,
-	RouteId extends AppRouteId | null = AppRouteId | null
-> {
-	/**
-	 * Parameters of the target page - e.g. for a route like `/blog/[slug]`, a `{ slug: string }` object.
-	 * Is `null` if the target is not part of the SvelteKit app (could not be resolved to a route).
-	 */
-	params: Params | null;
-	/**
-	 * Info about the target route
-	 */
-	route: {
-		/**
-		 * The ID of the current route - e.g. for `src/routes/blog/[slug]`, it would be `/blog/[slug]`. It is `null` when no route is matched.
-		 */
-		id: RouteId | null;
-	};
-	/**
-	 * The URL that is navigated to
-	 */
-	url: URL;
-	/**
-	 * The scroll position associated with this navigation.
-	 *
-	 * For the `from` target, this is the scroll position at the moment of navigation.
-	 *
-	 * For the `to` target, this represents the scroll position that will be or was restored:
-	 * - In `beforeNavigate` and `onNavigate`, this is only available for `popstate` navigations (back/forward button)
-	 *   and will be `null` for other navigation types, since the final scroll position isn't known
-	 *   ahead of time.
-	 * - In `afterNavigate`, this is always the scroll position that was applied after the navigation
-	 *   completed.
-	 */
-	scroll: { x: number; y: number } | null;
-}
-
-export interface GotoOptions {
-	/**
-	 * If `true`, replaces the current history entry rather than creating a new one.
-	 * @default false
-	 */
-	replace?: boolean;
-	/** @deprecated Use `replace` instead. */
-	replaceState?: boolean;
-	/**
-	 * If `true`, updates the URL and `page.state` without navigating.
-	 * @default false
-	 */
-	shallow?: boolean;
-	/**
-	 * If `true`, resets the scroll position (to the top of the page, or to the element
-	 * matching the URL's `#hash` if there is one) and resets focus (to the `<body>`, or the
-	 * `autofocus` element if there is one) once the navigation completes.
-	 *
-	 * If `false`, the current scroll position and focused element are left alone.
-	 * @default true, or false when `shallow` is true
-	 */
-	reset?: boolean;
-	/**
-	 * If `true`, reruns all `load` functions and queries of the page.
-	 * @default false
-	 */
-	refreshAll?: boolean;
-	/** Causes any `load` functions to rerun if they depend on one of the URLs. */
-	invalidate?: Array<string | URL | ((url: URL) => boolean)>;
-	/** @deprecated Use `refreshAll` instead. */
-	invalidateAll?: boolean;
-	/** An optional object that will be available as `page.state`. */
-	state?: App.PageState;
-	/**
-	 * If `true`, `page.state` will be restored after a full page reload.
-	 * @default false
-	 */
-	persistState?: boolean;
-}
-
-/**
- * - `enter`: The app has hydrated/started
- * - `form`: The user submitted a `<form method="GET">`
- * - `goto`: Navigation was triggered by a `goto(...)` call or a redirect
- * - `leave`: The app is being left either because the tab is being closed or a navigation to a different document is occurring
- * - `link`: Navigation was triggered by a link click
- * - `popstate`: Navigation was triggered by back/forward navigation
- */
-export type NavigationType = 'enter' | 'form' | 'leave' | 'link' | 'goto' | 'popstate';
-
-export interface NavigationBase {
-	/**
-	 * The type of navigation:
-	 * - `enter`: The app has hydrated/started
-	 * - `form`: The user submitted a `<form method="GET">`
-	 * - `goto`: Navigation was triggered by a `goto(...)` call or a redirect
-	 * - `leave`: The app is being left either because the tab is being closed or a navigation to a different document is occurring
-	 * - `link`: Navigation was triggered by a link click
-	 * - `popstate`: Navigation was triggered by back/forward navigation
-	 */
-	type: NavigationType;
-	/** Whether this is a shallow navigation. */
-	shallow: boolean;
-	/**
-	 * Where navigation was triggered from
-	 */
-	from: NavigationTarget | null;
-	/**
-	 * Where navigation is going to/has gone to
-	 */
-	to: NavigationTarget | null;
-	/**
-	 * Whether or not the navigation will result in the page being unloaded (i.e. not a client-side navigation).
-	 */
-	willUnload: boolean;
-	/**
-	 * A promise that resolves once the navigation is complete, and rejects if the navigation
-	 * fails or is aborted. In the case of a `willUnload` navigation, the promise will never resolve
-	 */
-	complete: Promise<void>;
-}
-
-/**
- * The navigation that occurs when the app starts/hydrates
- */
-export interface NavigationEnter extends NavigationBase {
-	type: 'enter';
-
-	/**
-	 * In case of a history back/forward navigation, the number of steps to go back/forward
-	 */
-	delta?: undefined;
-
-	/**
-	 * Dispatched `Event` object when navigation occurred by `popstate` or `link`.
-	 */
-	event?: undefined;
-}
-
-export type NavigationExternal = NavigationGoto | NavigationLeave;
-
-/**
- * A navigation triggered by a `goto(...)` call or a redirect
- */
-export interface NavigationGoto extends NavigationBase {
-	type: 'goto';
-}
-
-/**
- * A navigation triggered by the tab being closed, or the user navigating to a different document
- */
-export interface NavigationLeave extends NavigationBase {
-	type: 'leave';
-}
-
-/**
- * A navigation triggered by a `<form method="GET">`
- */
-export interface NavigationFormSubmit extends NavigationBase {
-	type: 'form';
-
-	/**
-	 * The `SubmitEvent` that caused the navigation
-	 */
-	event: SubmitEvent;
-}
-
-/**
- * A navigation triggered by back/forward navigation
- */
-export interface NavigationPopState extends NavigationBase {
-	type: 'popstate';
-
-	/**
-	 * In case of a history back/forward navigation, the number of steps to go back/forward
-	 */
-	delta: number;
-
-	/**
-	 * The `PopStateEvent` that caused the navigation
-	 */
-	event: PopStateEvent;
-}
-
-/**
- * A navigation triggered by a link click
- */
-export interface NavigationLink extends NavigationBase {
-	type: 'link';
-
-	/**
-	 * The `PointerEvent` that caused the navigation
-	 */
-	event: PointerEvent;
-}
-
-export type Navigation =
-	NavigationExternal | NavigationFormSubmit | NavigationPopState | NavigationLink;
-
-/**
- * The argument passed to [`beforeNavigate`](https://svelte.dev/docs/kit/$app-navigation#beforeNavigate) callbacks.
- */
-export type BeforeNavigate = Navigation & {
-	/**
-	 * Call this to prevent the navigation from starting.
-	 */
-	cancel: () => void;
-};
-
-/**
- * The argument passed to [`onNavigate`](https://svelte.dev/docs/kit/$app-navigation#onNavigate) callbacks.
- */
-export type OnNavigate = Navigation & {
-	type: Exclude<NavigationType, 'enter' | 'leave'>;
-	/**
-	 * Since `onNavigate` callbacks are called immediately before a client-side navigation, they will never be called with a navigation that unloads the page.
-	 */
-	willUnload: false;
-};
-
-/**
- * The argument passed to [`afterNavigate`](https://svelte.dev/docs/kit/$app-navigation#afterNavigate) callbacks.
- */
-export type AfterNavigate = (Navigation | NavigationEnter) & {
-	type: Exclude<NavigationType, 'leave'>;
-	/**
-	 * Since `afterNavigate` callbacks are called after a navigation completes, they will never be called with a navigation that unloads the page.
-	 */
-	willUnload: false;
-};
-
-/**
- * The shape of the [`page`](https://svelte.dev/docs/kit/$app-state#page) reactive object.
- */
-export interface Page<
-	Params extends AppLayoutParams<'/'> = AppLayoutParams<'/'>,
-	RouteId extends AppRouteId | null = AppRouteId | null
-> {
-	/**
-	 * The URL of the current page.
-	 */
-	url: ReadonlyURL & { readonly pathname: ResolvedPathname | (string & {}) };
-	/**
-	 * The parameters of the current page - e.g. for a route like `/blog/[slug]`, a `{ slug: string }` object.
-	 */
-	params: Params;
-	/**
-	 * Info about the current route.
-	 */
-	route: {
-		/**
-		 * The ID of the current route - e.g. for `src/routes/blog/[slug]`, it would be `/blog/[slug]`. It is `null` when no route is matched.
-		 */
-		id: RouteId;
-	};
-	/**
-	 * HTTP status code of the current page.
-	 */
-	status: number;
-	/**
-	 * The error object of the current page, if any. Filled from the `handleError` hooks.
-	 */
-	error: App.Error | null;
-	/**
-	 * The merged result of all data from all `load` functions on the current page. You can type a common denominator through `App.PageData`.
-	 */
-	data: App.PageData & Record<string, any>;
-	/**
-	 * The page state, which can be manipulated using [`goto`](https://svelte.dev/docs/kit/$app-navigation#goto) from `$app/navigation`.
-	 */
-	state: App.PageState;
-	/**
-	 * Information about the target of the current shallow navigation, or `null` if no shallow navigation has occurred.
-	 */
-	shallow: {
-		/** Parameters of the target route, or `null` if the URL does not resolve to a route. */
-		params: AppLayoutParams<'/'> | null;
-		/** Info about the target route, or `null` if the URL does not resolve to a route. */
-		route: { id: AppRouteId } | null;
-		/** The normalized URL passed to `goto(..., { shallow: true })`. */
-		url: ReadonlyURL;
-	} | null;
-	/**
-	 * Filled only after a form submission. See [form actions](https://svelte.dev/docs/kit/form-actions) for more info.
-	 */
-	form: any;
-}
-
-/**
  * The shape of a param matcher. See [matching](https://svelte.dev/docs/kit/advanced-routing#Matching) for more info.
  */
 export type ParamMatcher<Output = any> = StandardSchemaV1<string, Output>;
@@ -1933,30 +1641,6 @@ export type Actions<
 > = Record<string, Action<Params, OutputData, RouteId>>;
 
 /**
- * When calling a form action via fetch, the response will be one of these shapes.
- * ```svelte
- * <form method="post" use:enhance={() => {
- *   return ({ result }) => {
- * 		// result is of type ActionResult
- *   };
- * }}
- * ```
- *
- * Success and failure results carry the root-relative `pathname + search` of the action URL, with
- * the `?/actionName` parameter removed. Redirect results carry the redirect target. Server-generated
- * error results also carry the action location, while client-generated errors such as network
- * failures do not. `update` uses this location to emulate native form navigation.
- */
-export type ActionResult<
-	Success extends Record<string, unknown> | undefined = Record<string, any>,
-	Failure extends Record<string, unknown> | undefined = Record<string, any>
-> =
-	| { type: 'success'; status: number; data?: Success; location: string }
-	| { type: 'failure'; status: number; data?: Failure; location: string }
-	| { type: 'redirect'; status: number; location: string }
-	| { type: 'error'; status?: number; error: App.Error; location?: string };
-
-/**
  * The object returned by the [`error`](https://svelte.dev/docs/kit/@sveltejs-kit#error) function.
  */
 export interface HttpError {
@@ -1975,37 +1659,6 @@ export interface Redirect {
 	/** The location to redirect to. */
 	location: string;
 }
-
-export type SubmitFunction<
-	Success extends Record<string, unknown> | undefined = Record<string, any>,
-	Failure extends Record<string, unknown> | undefined = Record<string, any>
-> = (input: {
-	action: URL;
-	formData: FormData;
-	formElement: HTMLFormElement;
-	controller: AbortController;
-	submitter: HTMLElement | null;
-	cancel: () => void;
-}) => MaybePromise<
-	| void
-	| ((opts: {
-			formData: FormData;
-			formElement: HTMLFormElement;
-			action: URL;
-			result: ActionResult<Success, Failure>;
-			/**
-			 * Call this to get the default behavior of a form submission response.
-			 * @param options Set `reset: false` if you don't want the `<form>` values to be reset after a successful submission. `refreshAll` defaults to `true` for successful results and `false` for failures. When the submission navigates, setting it to `false` still runs the destination's `load` functions but may reuse shared layout data. Set `navigate: false` to apply non-redirect results to the current page instead of navigating to `result.location`. Redirects are always followed.
-			 */
-			update: (options?: {
-				reset?: boolean;
-				refreshAll?: boolean;
-				navigate?: boolean;
-				/** @deprecated Use `refreshAll` instead. */
-				invalidateAll?: boolean;
-			}) => Promise<void>;
-	  }) => MaybePromise<void>)
->;
 
 /**
  * The type of `export const snapshot` exported from a page or layout component.
