+++ packages/kit/src/runtime/app/stores.js
@@ -14,7 +14,7 @@ export const getStores = removed;
 
 /**
  * @deprecated Use `page` from `$app/state` instead (requires Svelte 5, [see docs for more info](https://svelte.dev/docs/kit/migrating-to-sveltekit-2#SvelteKit-2.12:-$app-stores-deprecated))
- * @type {import('svelte/store').Readable<import('@sveltejs/kit').Page>}
+ * @type {import('svelte/store').Readable<import('$app/state').Page>}
  */
 export const page = {
 	subscribe: removed
@@ -22,7 +22,7 @@ export const page = {
 
 /**
  * @deprecated Use `navigating` from `$app/state` instead (requires Svelte 5, [see docs for more info](https://svelte.dev/docs/kit/migrating-to-sveltekit-2#SvelteKit-2.12:-$app-stores-deprecated))
- * @type {import('svelte/store').Readable<import('@sveltejs/kit').Navigation | null>}
+ * @type {import('svelte/store').Readable<import('$app/navigation').Navigation | null>}
  */
 export const navigating = {
 	subscribe: removed
