+++ packages/kit/test/apps/basics/src/routes/navigation-lifecycle/before-navigate/prevent-navigation/+page.svelte
@@ -3,7 +3,7 @@
 
 	let times_triggered = 0;
 	let unload = false;
-	/** @type {import("@sveltejs/kit").BeforeNavigate['type']} */
+	/** @type {import("$app/navigation").BeforeNavigate['type']} */
 	let navigation_type;
 	beforeNavigate(({ cancel, type, willUnload, to }) => {
 		times_triggered++;
