+++ packages/kit/test/types.d.ts
@@ -8,7 +8,7 @@ import {
 } from '@playwright/test';
 import { IncomingMessage, ServerResponse } from 'node:http';
 import '../types/index.d.ts';
-import { AfterNavigate, BeforeNavigate, GotoOptions } from '@sveltejs/kit';
+import { AfterNavigate, BeforeNavigate, GotoOptions } from '$app/navigation';
 
 export const test: TestType<
 	Omit<PlaywrightTestArgs, 'page'> &
