+++ packages/kit/tsconfig.json
@@ -15,10 +15,13 @@
 			"@sveltejs/kit/internal/server": ["./src/exports/internal/server/index.js"],
 			"$app/env": ["./src/runtime/app/env/types.d.ts"],
 			"$app/env/internal": ["./src/runtime/app/env/internal.js"],
+			"$app/forms": ["./src/runtime/app/forms/public.d.ts"],
+			"$app/navigation": ["./src/runtime/app/navigation/public.d.ts"],
 			"$app/paths": ["./src/runtime/app/paths/public.d.ts"],
 			"$app/paths/internal/client": ["./src/runtime/app/paths/internal/client.js"],
 			"$app/paths/internal/server": ["./src/runtime/app/paths/internal/server.js"],
 			"$app/server": ["./src/runtime/app/server/index.js"],
+			"$app/state": ["./src/runtime/app/state/public.d.ts"],
 			// internal use only
 			"types": ["./src/types/internal.d.ts"]
 		},
